To the bulk of the SettingsDialog template code and broke it into individual configuration section components.
These new components are located in the configuration-sections folder
