+++ src/components/configuration-sections/WebLlmConfigSection.vue
@@ -1,0 +1,59 @@
+<template>
+    <div class="control select-dropdown">
+        <label for="localModelsSelection">Model To Load In Browser:</label>
+        <select id="localModelsSelection" :value="browserModelSelection"
+            @change="handleUpdate('browserModelSelection', $event.target.value)">
+            <option value="Llama-3-8B-Instruct-q4f32_1">Llama-3-8B-Instruct-q4f32 (~6.1gb VRAM)</option>
+            <option value="Llama-3-8B-Instruct-q4f16_1-1k">Llama-3-8B-Instruct-q4f16 1k Context (~4.6gb VRAM)
+            </option>
+            <option value="Llama-3-8B-Instruct-q4f32_1-1k">Llama-3-8B-Instruct-q4f32 1k Context (~5.2gb VRAM)
+            </option>
+            <option value="Llama-2-7b-chat-hf-q4f16_1">Llama-2-7b-chat-hf-q4f16 (~6.8gb VRAM)</option>
+            <option value="TinyLlama-1.1B-Chat-v0.4-q4f32_1-1k">TinyLlama-1.1B-Chat-v0.4-q4f32 1k Context (~1.0gb
+                VRAM)</option>
+            <option value="TinyLlama-1.1B-Chat-v0.4-q0f32">TinyLlama-1.1B-Chat-v0.4-q0f32 (~5.3gb VRAM)</option>
+            <option value="Mistral-7B-Instruct-v0.2-q4f16_1">Mistral-7B-Instruct-v0.2 (~6.1gb VRAM)</option>
+            <option value="OpenHermes-2.5-Mistral-7B-q4f16_1">OpenHermes-2.5-Mistral-7B (~6.1gb VRAM)</option>
+            <option value="WizardMath-7B-V1.1-q4f16_1">WizardMath-7B-V1.1-q4f16 (~6.1gb VRAM)</option>
+            <option value="NeuralHermes-2.5-Mistral-7B-q4f16_1">NeuralHermes-2.5-Mistral-7B-q4f16 (~6.1gb VRAM)
+            </option>
+            <option value="gemma-2b-it-q4f32_1">gemma-2b-it-q4f32 (~1.8gb VRAM)</option>
+            <option value="gemma-2b-it-q4f32_1-1k">gemma-2b-it-q4f32 1k Context (~1.6gb VRAM)</option>
+        </select>
+    </div>
+</template>
+
+<script setup>
+import { ref } from 'vue';
+import { browserModelSelection } from '@/libs/state-management/state';
+import { handleUpdate } from '@/libs/utils/settings-utils';
+</script>
+
+<style scoped>
+.select-dropdown select {
+    appearance: none;
+    background-color: #333;
+    color: #fff;
+    max-width: 65vw;
+    height: 40px;
+    width: 100%;
+    padding-left: 6px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    font-size: 16px;
+
+    &:hover {
+        background-color: #444;
+    }
+
+    &:focus {
+        outline: none;
+    }
+}
+
+.select-dropdown option {
+    background-color: #222;
+    color: #fff;
+}
+</style>
