+++ src/libs/utils/settings-utils.js
@@ -211,3 +211,49 @@ export function importSettings(settingsData, update) {
     }
   }
 }
+
+export function handleDeleteSystemPrompt(index) {
+  deleteSystemPrompt(index, showToast);
+}
+
+export function handleSelectSystemPrompt(index) {
+  selectSystemPrompt(index, systemPrompt);
+}
+
+// Custom Configs
+export function handleDeleteCustomConfig(index) {
+  deleteCustomConfig(index, showToast);
+}
+
+export function handleUpdate(field, value) {
+  update(field, value);
+}
+
+
+export function updateGptSliderValue(value) {
+  handleUpdate('sliderValue', parseFloat(value));
+}
+
+export function updateWhisperSlider(value) {
+  handleUpdate('whisper-temperature', parseFloat(value));
+}
+
+export function updateLocalSliderValue(value) {
+  handleUpdate('localSliderValue', parseFloat(value));
+}
+
+export function updateClaudeSliderValue(value) {
+  handleUpdate('claudeSliderValue', parseFloat(value));
+}
+
+export function updateTopPSliderValue(value) {
+  handleUpdate('top_P', parseFloat(value));
+}
+
+export function updateRepetitionSliderValue(value) {
+  handleUpdate('repetitionPenalty', parseFloat(value));
+}
+
+export function handleSelectCustomConfig(index) {
+  selectCustomConfig(index, localModelEndpoint, localModelKey, localModelName, maxTokens, localSliderValue, top_P, repetitionPenalty);
+}
